+++ tag/badgenametag/badgenametag.go
@@ -1,0 +1,1 @@
+package badgenametag
