+++ tag/bolt11tag/bolt11tag.go
@@ -1,0 +1,1 @@
+package bolt11tag
