+++ tag/delegationtag/contentwarntag.go
@@ -1,0 +1,1 @@
+package delegationtag
