+++ tag/eventidtag/eventidtag.go
@@ -1,0 +1,1 @@
+package eventidtag
