+++ tag/eventtag/eventtag.go
@@ -1,0 +1,1 @@
+package eventtag
