+++ tag/hashtagtag/hashtagtag.go
@@ -1,0 +1,1 @@
+package hashtagtag
