+++ tag/lnurltag/lnurltag.go
@@ -1,0 +1,1 @@
+package lnurltag
