+++ tag/preimagetag/preimagetag.go
@@ -1,0 +1,1 @@
+package preimagetag
