+++ tag/pubkeytag/pubkeytag.go
@@ -1,0 +1,1 @@
+package pubkeytag
