+++ tag/publishedattag/publishedattag.go
@@ -1,0 +1,1 @@
+package publishedattag
