+++ tag/thumbtag/thumbtag.go
@@ -1,0 +1,1 @@
+package thumbtag
